Sweet, got it working
